reformat helper and common scripts
